+++ crates/interpreter/src/constants/abis.rs
@@ -15,31 +15,8 @@
 
 #![allow(clippy::expect_used)]
 
-use crate::{
-    adapter::Adapter,
-    adapters::transfer::{
-        erc1155::ERC1155Adapter, erc20::ERC20Adapter, erc721::ERC721Adapter, eth::EthAdapter,
-    },
-};
-use ethers_main::{abi::parse_abi, contract::BaseContract, types::H256};
+use ethers_main::{abi::parse_abi, contract::BaseContract};
 use lazy_static::lazy_static;
-use serde::{Deserialize, Serialize};
-use std::str::FromStr;
-use strum_macros::EnumVariantNames;
-
-lazy_static! {
-    #[derive(Clone)]
-    pub static ref ADAPTERS: Vec<Box<dyn Adapter + Sync + Send>> =
-        vec![Box::new(EthAdapter::new()), Box::new(ERC20Adapter::new()), Box::new(ERC721Adapter::new()), Box::new(ERC1155Adapter::new())];
-}
-
-lazy_static! {
-    pub static ref TRANSFER_EVENT_TOPIC: H256 =
-        // https://www.4byte.directory/event-signatures/?bytes_signature=0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef
-        // Transfer(address,address,uint256)
-        H256::from_str("0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef")
-            .expect("Failed to parse address");
-}
 
 lazy_static! {
     pub static ref ERC20_ABI: BaseContract = BaseContract::from(
@@ -72,24 +49,4 @@ lazy_static! {
         ])
         .expect("Failed to parse ABI"),
     );
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, EnumVariantNames)]
-pub enum InterpretationActionType {
-    #[strum(serialize = "NATIVE_RECEIVE")]
-    NativeReceive,
-    #[strum(serialize = "NATIVE_SEND")]
-    NativeSend,
-    #[strum(serialize = "ERC20_RECEIVE")]
-    ERC20Receive,
-    #[strum(serialize = "ERC20_SEND")]
-    ERC20Send,
-    #[strum(serialize = "ERC721_RECEIVE")]
-    ERC721Receive,
-    #[strum(serialize = "ERC721_SEND")]
-    ERC721Send,
-    #[strum(serialize = "ERC721_MINTED")]
-    ERC721Minted,
-    #[strum(serialize = "ERC721_BURNED")]
-    ERC721Burned,
 }
